refactor: removed most linter warnings and imported the Riverpod package for state management.
