+++ lib/presentation/widgets/contact_section/contact_section.dart
@@ -126,7 +126,7 @@ class ContactSection extends StatelessWidget {
   }
 
   Row buildNameEmailFieldDesktop() {
-    return Row(
+    return const Row(
       children: [
         Flexible(
           child: CustomTextField(
@@ -146,7 +146,7 @@ class ContactSection extends StatelessWidget {
   }
 
   Column buildNameEmailFieldMobile() {
-    return Column(
+    return const Column(
       children: [
         Flexible(
           child: CustomTextField(
