+++ lib/presentation/widgets/drawer_mobile.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:web_portfolio/config/constants/colors.dart';
-import 'package:web_portfolio/config/constants/nav-items.dart';
+import 'package:web_portfolio/config/constants/nav_items.dart';
 
 class DrawerMobile extends StatelessWidget {
   const DrawerMobile({super.key, required this.onNavItemTap});
