+++ lib/presentation/widgets/header_section/header_web_desktop.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:web_portfolio/config/constants/colors.dart';
-import 'package:web_portfolio/config/constants/nav-items.dart';
+import 'package:web_portfolio/config/constants/nav_items.dart';
 import 'package:web_portfolio/presentation/widgets/site_logo.dart';
 
 class HeaderDesktop extends StatelessWidget {
