+++ pubspec.yaml
@@ -3,27 +3,24 @@ description: "A new Flutter project."
 
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-
 version: 1.0.0+1
 
 environment:
   sdk: '>=3.4.3 <4.0.0'
 
-
 dependencies:
+  cupertino_icons: ^1.0.6
   flutter:
     sdk: flutter
-
-
-
-  cupertino_icons: ^1.0.6
+  flutter_riverpod: ^2.5.1
 
 dev_dependencies:
+
+  custom_lint: ^0.6.4
+  flutter_lints: ^3.0.0
   flutter_test:
     sdk: flutter
-
-  flutter_lints: ^3.0.0
-
+  riverpod_lint: ^2.3.10
 
 flutter:
